move serduszko system to serduszko module
