+++ plugin/pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>pl.mgtm</groupId>
     <artifactId>MagicznaKraina</artifactId>
-    <version>1.0.0</version>
+    <version>1.0.1</version>
     <packaging>jar</packaging>
 
     <name>MagicznaKraina</name>
